requirements.txt
# Core dependencies
python-telegram-bot==20.7
python-dotenv==1.0.0
asyncio==3.4.3

# Database
SQLAlchemy==2.0.23
aiosqlite==0.19.0

# Utilities
schedule==1.2.0
pytz==2023.3

# Logging and monitoring
colorlog==6.8.0

# Data handling
pydantic==2.5.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.12.0
flake8==7.0.0
pylint==3.0.3